move tests to proxy_test
